# openapi.yaml
openapi: 3.0.2
info:
  title: API de Artesanatos da Maya
  description: Sistema completo de gestão de produtos, materiais e comentários
  version: 2.0.0

servers:
  - url: https://api-one-delta-89.vercel.app
    description: API em produção
  - url: http://localhost:5000
    description: Servidor local

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /:
    get:
      summary: Boas-vindas
      responses:
        200:
          description: Sucesso

  /login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario:
                  type: string
                senha:
                  type: string
      responses:
        200:
          description: Login realizado
        401:
          description: Credenciais inválidas

  /produtos:
    get:
      summary: Listar produtos
      responses:
        200:
          description: Lista de produtos

  /criarprodutos:
    post:
      summary: Criar produto
      security:
        - BearerAuth: []
      responses:
        201:
          description: Produto criado

  /produtos/{id}:
    get:
      summary: Buscar produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Produto encontrado
        404:
          description: Não encontrado
    put:
      summary: Atualizar produto
      security:
        - BearerAuth: []
      responses:
        200:
          description: Produto atualizado
    patch:
      summary: Atualizar parcial
      security:
        - BearerAuth: []
      responses:
        200:
          description: Produto atualizado
    delete:
      summary: Excluir produto
      security:
        - BearerAuth: []
      responses:
        200:
          description: Produto excluído

  /materiais:
    get:
      summary: Listar materiais
      security:
        - BearerAuth: []
      responses:
        200:
          description: Lista de materiais
    post:
      summary: Criar material
      security:
        - BearerAuth: []
      responses:
        201:
          description: Material criado

  /materiais/{id}:
    get:
      summary: Buscar material
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Material encontrado
    put:
      summary: Atualizar material
      security:
        - BearerAuth: []
      responses:
        200:
          description: Material atualizado
    patch:
      summary: Atualizar parcial
      security:
        - BearerAuth: []
      responses:
        200:
          description: Material atualizado
    delete:
      summary: Excluir material
      security:
        - BearerAuth: []
      responses:
        200:
          description: Material excluído

  /comentarios/{produto_id}:
    get:
      summary: Listar comentários de um produto
      parameters:
        - name: produto_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Lista de comentários

  /comentarios:
    post:
      summary: Adicionar um comentário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                produto_id:
                  type: integer
                nome:
                  type: string
                estrelas:
                  type: integer
                  minimum: 1
                  maximum: 5
                texto:
                  type: string
      responses:
        201:
          description: Comentário adicionado
        400:
          description: Dados inválidos

  /admin/comentarios:
    get:
      summary: Listar todos os comentários (admin)
      security:
        - BearerAuth: []
      responses:
        200:
          description: Lista de todos os comentários

  /admin/comentarios/{comentario_id}:
    delete:
      summary: Excluir comentário (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: comentario_id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Comentário excluído
        404:
          description: Comentário não encontrado
    /pedidos:
    get:
      summary: Listar todos os pedidos (admin)
      security:
        - BearerAuth: []
      responses:
        200:
          description: Lista de pedidos
    post:
      summary: Criar um novo pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itens:
                  type: string
                total:
                  type: number
                cliente:
                  type: string
                whatsapp:
                  type: string
      responses:
        201:
          description: Pedido salvo

  /admin/pedidos/{pedido_id}:
    delete:
      summary: Excluir pedido (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: pedido_id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Pedido excluído

  /admin/pedidos/{pedido_id}/status:
    patch:
      summary: Atualizar status do pedido (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: pedido_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        200:
          description: Status atualizado